Isolate the DNS localhost tests
