+++ resolver-dns/src/main/java/io/netty/resolver/dns/DnsNameResolver.java
@@ -65,6 +65,7 @@ import io.netty.util.internal.StringUtil;
 import io.netty.util.internal.SystemPropertyUtil;
 import io.netty.util.internal.logging.InternalLogger;
 import io.netty.util.internal.logging.InternalLoggerFactory;
+import org.jetbrains.annotations.VisibleForTesting;
 
 import java.lang.reflect.Method;
 import java.net.IDN;
@@ -148,11 +149,8 @@ public class DnsNameResolver extends InetNameResolver {
             "io.netty.resolver.dns.resolveLocalhostWithoutDns";
 
     /** The runtime value of {@link #RESOLVE_LOCALHOST_WITHOUT_DNS_PROPERTY}. */
-    private static boolean resolveLocalhostWithoutDns;
-
-    static void setResolveLocalhostWithoutDns(final boolean value) {
-        resolveLocalhostWithoutDns = value;
-    }
+    @VisibleForTesting
+    static boolean resolveLocalhostWithoutDns;
 
     static final ResolvedAddressTypes DEFAULT_RESOLVE_ADDRESS_TYPES;
     static final String[] DEFAULT_SEARCH_DOMAINS;
